# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Gym Studio SaaS API
  version: 0.1.0
  description: >
    REST API for managing gym memberships, billing, classes, attendance, and access control.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:4000
    description: Development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
paths:
...
  /api/members:
    get:
      summary: List members
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Member collection
    post:
      summary: Create member
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Member created
...